CONPY-189: Windows build fix for Visual Studio 2022
- Build parameters are now written to config_win.h instead of passing
  them as command line parameters
- Fixed several compiler warnings

+++ mariadb/mariadb_cursor.c
@@ -1102,7 +1102,7 @@ MrdbCursor_execute_text(MrdbCursor *self, PyObject *args)
     db= self->connection->mysql;
 
     Py_BEGIN_ALLOW_THREADS;
-    rc= mysql_send_query(db, statement, statement_len);
+    rc= mysql_send_query(db, statement, (long)statement_len);
     Py_END_ALLOW_THREADS;
 
     if (rc)
